Batch 19: Audio/gamelogic/control_bar/terrain build fixes — 32 files
- Audio: rubato SincInterpolationParameters, rodio f32 conversion, SpatialSink Debug, cpal thread safety
- Gamelogic: player.rs crc mutability (15 fields), ai_player.rs Bool→bool, stealth_upgrade type fixes, pathfinder stubs, contain module xfer stubs, slowdeathbehavior type fixes
- control_bar.rs: added missing fn update_context_observer header (orphaned code at depth=1 inside impl)
- terrain_visual.rs: removed extra closing brace
- game_engine + gamelogic crates compile clean (0 errors)

+++ GeneralsRust/Code/GameEngine/Common/src/common/thing/thing.rs
@@ -236,7 +236,7 @@ impl Thing for BaseThing {
     }
 
     fn is_kind_of(&self, kind: KindOfType) -> bool {
-        self.template.is_kind_of(kind)
+        self.template.is_kind_of(kind as u64)
     }
 
     fn is_kind_of_multi(
